Set scheduler algorithm provider as ClusterAutoscalerProvider
The default scheduler algorithm provider uses least-requested algorithm, which selects nodes based on the least resource usage. That's to say, it tends to deploy pods onto the biggest machine first. This is not what we want. We would like to deploy pods requesting less resources onto smaller machines. Therefore, we change to config to use the most-requested algorithm.